+++ R/cookies.R
@@ -71,6 +71,7 @@ storeCookies <- function(requestURL, cookieHeaders) {
   lapply(cookies, function(co) {
     # Remove any duplicates
     # RFC says duplicate cookies are ones that have the same domain, name, and path
+    # nolint start
     hostCookies <<- hostCookies[
       !(co$name == hostCookies$name & co$path == hostCookies$path),
     ]
@@ -80,6 +81,7 @@ storeCookies <- function(requestURL, cookieHeaders) {
       as.data.frame(co, stringsAsFactors = FALSE),
       hostCookies
     )
+    # nolint end
   })
 
   # Save this host's cookies into the cookies store.